pyproject.toml
[project]
name = "openslides"
version = "2.0.0"
description = "Prompt + URL -> branded pitch deck. A floom app."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "anthropic==0.49.0",
    "playwright==1.51.0",
    "httpx==0.28.1",
    "Pillow==11.1.0",
    "PyPDF2==3.0.1",
    "python-pptx==1.0.2",
    "colorthief==0.2.1",
]

[project.optional-dependencies]
audit = ["google-genai==1.9.0"]
all = ["openslides[audit]"]

[project.scripts]
openslides = "openslides.main:cli"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
